Refactored functions around
